third_party/WebKit/LayoutTests/accessibility/aria-checkbox-checked-expected.txt

 
X
X
   
This tests that ARIA checkboxes correctly handle the aria-checked attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS checkbox.checkedState is "false"
PASS checkbox.checkedState is "true"
PASS checkbox.checkedState is "false"
PASS checkbox.checkedState is "true"
PASS checkbox.checkedState is "false"
PASS checkbox.checkedState is "false"
PASS checkbox.checkedState is "true"
PASS checkbox.checkedState is "true"
PASS successfullyParsed is true

TEST COMPLETE
